+++ api/operator/v1/vlcluster_types.go
@@ -899,18 +899,6 @@ func (cr *VLCluster) StorageURL() string {
 	return fmt.Sprintf("%s://%s.%s.svc:%s", vmv1beta1.HTTPProtoFromFlags(cr.Spec.VLStorage.ExtraArgs), cr.GetVLStorageName(), cr.Namespace, port)
 }
 
-func (cr *VLCluster) AsComponentURL(component string) string {
-	switch component {
-	case "vlinsert":
-		return cr.InsertURL()
-	case "vlselect":
-		return cr.SelectURL()
-	case "vlstorage":
-		return cr.StorageURL()
-	}
-	return ""
-}
-
 // +kubebuilder:object:root=true
 
 // VLClusterList contains a list of VLCluster