+++ api/operator/v1beta1/vmsingle_types.go
@@ -310,10 +310,6 @@ func (cr *VMSingle) AsURL() string {
 	return fmt.Sprintf("%s://%s.%s.svc:%s", HTTPProtoFromFlags(cr.Spec.ExtraArgs), cr.PrefixedName(), cr.Namespace, port)
 }
 
-func (cr *VMSingle) AsComponentURL(_ string) string {
-	return cr.AsURL()
-}
-
 // LastAppliedSpecAsPatch return last applied single spec as patch annotation
 func (cr *VMSingle) LastAppliedSpecAsPatch() (client.Patch, error) {
 	return LastAppliedChangesAsPatch(cr.ObjectMeta, cr.Spec)